refactor(racing): Game을 service 패키지에서 domain 패키지로 이동
